Add files via upload
Write a program inC or C++ to implement an integer adder that can be interrupted by a SIGINT (i.e. Ctrl-C) signal. 

our program should first prompt the    user for a large positive integer (e.g. 10000000). Once the integer is entered by the user, your program should then keep adding integers from 1 until the integer entered is reached. Your program should NOT be interrupted by any signals (including SIGINT) when it is waiting for the integer to be entered by the user. However, it can only be interrupted by an SIGINT signal when it is in the process of performing the addition operation. When anSIGINT signal is caught, it should be handled by a signal handler that will print (to the screen) whatever has been done at the time when the program was interrupted by the signal and then terminate the program. Note that if the addition operation is completed without being interrupted by an SIGINT signal, your program should repeat the same process by prompting the user for a new large integer and calculate the summation for the integer. Be sure to reinitialize your "summation" variable to 0 before the process starts over.